Move shards package to root and rename to 'search'
