+++ search/shards_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package shards
+package search
 
 import (
 	"bytes"
